requirements.txt
aiogram==3.15.0
aiosqlite==0.20.0
python-dotenv==1.0.1
pydantic-settings==2.7.0
rapidfuzz==3.10.1
beautifulsoup4==4.12.3
lxml==5.3.0
httpx==0.28.1
